make .dict.dz the default target
The previous default target built a release. That is not what a user
expects when typing "make", he/she expects a build. A separate "release"
target has been introduced.

At the moment, the dictd platform is the only supported. For the future,
it'd be nicer to have a proper build target depending on the specific
platforms and those in turn on the files.