Remove dict-tgz links and platform
It doesn't make sense to define dict.tgz and dict.bz2 as two separate
platforms, both contain the same files. Both formats can be extracted on
virtually all systems, so it is best to take the more modern one and
drop the older.
This attempts to remove all references to dict-tgz,  while not breaking
the web site.